refactor: extract shared match naming and achievement helpers
Entire-Checkpoint: 01M005KFSFHZ69A9RFQSQJ0791
