+++ apps/web/src/lib/achievements.ts
@@ -1,0 +1,3 @@
+export function formatAchievementName(type: string): string {
+	return type.replace(/_/g, " ").replace(/\b\w/g, (char) => char.toUpperCase());
+}
